feat(multiple mappings): add more than one mapping during boot (Ubuntu)
rapiddisk-on-boot is able to create many ramdisks and map them to
different devices during boot. This is useful to cache volumes mounted
during boot and to create a wb+wa cache (Ubuntu).
Fix a bug in rapiddisk_clean in a regular expression, which caused the
script to try to remove all the ramdisks and not just the unused ones.
(Ubuntu).
rapiddisk-on-boot uninstalls each configured mapping via an interactive
menu. (Ubuntu)